backend/python/fastapi_graphql_mongo/requirements.txt
fastapi
uvicorn
strawberry-graphql
pymongo
motor
python-dotenv
beanie
